Fix up --polar day view     Revive aged polar view code     Add sample --polar day view to images/     Collapse common title-drawing code into common function
Note that the --polar view is different from year and cartesian day view.
The polar view takes into account that mean time hour angles and solar
position azimuth angles may be wildly different. The polar view is
essentially an azimuth/elevation plot of the sun's position for the day.

 * The azimuth angle is plotted from the center of the circle with north
   being straight down. This gives a view with noon at the top of the plot
   and makes more sense for observers in the northern hemisphere. [1]
 * The solar elevation is plotted with the nadir at the center of the
   circle, the horizon as the boundary between the twilight color and the
   whitish daylight color, and the zenith at the outer edge of the circle.

This view satisfies some of my research goals but probably isn't that
interesting for most folks.

[1] For observers in the southern hemisphere I could see having the azimuth
angles increase counterclockwise instead of clockwise to more instinctively
reflect how the sun traverses the sky in daily life. Support for observer
latitudes in the southern hemisphere is TBD, TODO, and unlikely anytime soon.